supervisor/supervisor-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supervisor-agent"
version = "1.0.0"
description = "Supervisor agent for coordinating large-scale workflows"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Arabic Themes Ecosystem", email = "support@yourdomain.com"}
]
dependencies = [
    "langgraph>=0.2.0",
    "fastapi>=0.110.0",
    "psycopg2-binary",
    "redis[hiredis]",
    "resend",
    "apscheduler",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
